Combine manifest tool files
